backend/requirements.txt
aiogram            ==  3.15.0
aiofiles           ==  24.1.0
aiohappyeyeballs   ==  2.4.3
aiohttp            ==  3.10.11
aiosignal          ==  1.3.1
alembic            ==  1.14.0
annotated-types    ==  0.7.0
asyncpg            ==  0.30.0
attrs              ==  24.2.0
certifi            ==  2024.8.30
charset-normalizer ==  3.4.1
frozenlist         ==  1.5.0
greenlet           ==  3.1.1
idna               ==  3.10
iniconfig          ==  2.0.0
magic-filter       ==  1.0.12
Mako               ==  1.3.6
MarkupSafe         ==  3.0.2
multidict          ==  6.1.0
packaging          ==  24.2
pip                ==  25.0.1
pluggy             ==  1.5.0
propcache          ==  0.2.0
psycopg            ==  3.2.4
psycopg2-binary    ==  2.9.10
pydantic           ==  2.9.2
pydantic_core      ==  2.23.4
pytest             ==  8.3.3
pytest-asyncio     ==  0.25.2
pytest-dotenv      ==  0.5.2
python-dotenv      ==  1.0.1
pythontest         ==  0.1.2
requests           ==  2.32.3
SQLAlchemy         ==  2.0.36
typing_extensions  ==  4.12.2
urllib3            ==  2.3.0
yarl               ==  1.18.0